Added chapter reset logic and refactor code

+++ src/main/kotlin/core/mangacollector/service/Collector.kt
@@ -5,6 +5,7 @@ import core.mangacollector.model.LatestMangaUpdate
 import core.mangacollector.model.Manga
 import core.mangacollector.model.UpdateStatus
 import core.mangacollector.repository.LatestUpdateRepository
+import core.mangacollector.repository.MangaCompactRepository
 import core.mangacollector.repository.MangaRepository
 import core.mangacollector.repository.UpdateStatusRepository
 import org.jsoup.Jsoup
@@ -22,7 +23,8 @@ import kotlin.collections.LinkedHashSet
 @Service
 class Collector(val luRepo: LatestUpdateRepository,
                 val mangaRepository: MangaRepository,
-                val updateStatusRepository: UpdateStatusRepository) {
+                val updateStatusRepository: UpdateStatusRepository,
+                val mangaCompactRepository: MangaCompactRepository) {
     val logger = LoggerFactory.getLogger(Collector::class.java)
 
     @Value("\${collector.url.fixer.full.start}")
@@ -34,35 +36,37 @@ class Collector(val luRepo: LatestUpdateRepository,
     }
 
     fun brokenUrlFixer() {
-        val page = 1;
-        val pageItem = 10;
-        var updates = updateStatusRepository.findAll(PageRequest.of(page, pageItem));
+        val page = 1
+        val pageItem = 10
+        var updates = updateStatusRepository.findAll(PageRequest.of(page, pageItem))
         logger.info("Performing batch url fixes for ${updates.totalPages} pages for ${updates.totalElements} items")
         for (x in 1..updates.totalPages) {
             logger.info("Performing url fixes for batch : $x")
-            updates = updateStatusRepository.findAll(PageRequest.of(x, pageItem));
+            updates = updateStatusRepository.findAll(PageRequest.of(x, pageItem))
             updates.content.forEach { update ->
-                val manga = mangaRepository.findByMangaUrl(update.mangaUrl)
-                manga?.let {
-                    logger.info("Fixing broken chapter for manga : ${manga.mangaName}")
-                    val updatedChapters = LinkedHashSet<Chapter>()
-                    val chapters = manga.chapters
-                    for (chapter in chapters) {
-                        val updatedImages = LinkedHashSet<String>()
-                        val images = chapter.chapterImages
-                        for (image in images) {
-                            var updated = image.replace(Regex("(s\\d+\\.m)"), "s8.m").replace(Regex("(\\d+\\.com)"), "8.com")
-                            updated = updated.replace("https://bu.mkklcdnbuv8.com", "https://bu.mkklcdnbuv1.com")
-                            updatedImages.add(updated)
+                val mangas = mangaRepository.findByMangaUrl(update.mangaUrl)
+                if (mangas != null && mangas.isNotEmpty()) {
+                    mangas.first().let { manga ->
+                        logger.info("Fixing broken chapter for manga : ${manga.mangaName}")
+                        val updatedChapters = LinkedHashSet<Chapter>()
+                        val chapters = manga.chapters
+                        for (chapter in chapters) {
+                            val updatedImages = LinkedHashSet<String>()
+                            val images = chapter.chapterImages
+                            for (image in images) {
+                                var updated = image.replace(Regex("(s\\d+\\.m)"), "s8.m").replace(Regex("(\\d+\\.com)"), "8.com")
+                                updated = updated.replace("https://bu.mkklcdnbuv8.com", "https://bu.mkklcdnbuv1.com")
+                                updatedImages.add(updated)
+                            }
+                            chapter.chapterImages = updatedImages
+                            updatedChapters.add(chapter)
                         }
-                        chapter.chapterImages = updatedImages
-                        updatedChapters.add(chapter)
+                        manga.chapters = updatedChapters
+                        mangaRepository.save(manga)
+                        logger.info("Removing updates from table : ${update}")
+                        updateStatusRepository.delete(update)
                     }
-                    manga.chapters = updatedChapters
-                    mangaRepository.save(manga)
-                    logger.info("Removing updates from table : ${update}")
-                    updateStatusRepository.delete(update)
-                } ?: run {
+                } else {
                     logger.info("No manga found to fix url with name : ${update.mangaUrl}")
                 }
             }
@@ -71,15 +75,15 @@ class Collector(val luRepo: LatestUpdateRepository,
     }
 
     fun brokenUrlFixerFull() {
-        val pageItem = 10;
-        var updates = mangaRepository.findAll(PageRequest.of(urlFixerStartPage, pageItem));
+        val pageItem = 10
+        var updates = mangaRepository.findAll(PageRequest.of(urlFixerStartPage, pageItem))
         logger.info("Performing batch url fixes for ${updates.totalPages} pages for ${updates.totalElements} items")
         for (x in urlFixerStartPage..updates.totalPages) {
             logger.info("Performing url fixes for batch : $x")
-            updates = mangaRepository.findAll(PageRequest.of(x, pageItem));
+            updates = mangaRepository.findAll(PageRequest.of(x, pageItem))
             updates.content.forEach { update ->
-                val manga = mangaRepository.findByMangaUrl(update.mangaUrl)
-                manga?.let {
+                val mangas = mangaRepository.findByMangaUrl(update.mangaUrl)
+                if (mangas != null && mangas.isNotEmpty()) mangas.first().let { manga ->
                     logger.info("Fixing broken chapter for manga : ${manga.mangaName}")
                     val updatedChapters = LinkedHashSet<Chapter>()
                     val chapters = manga.chapters
@@ -96,7 +100,7 @@ class Collector(val luRepo: LatestUpdateRepository,
                     }
                     manga.chapters = updatedChapters
                     mangaRepository.save(manga)
-                } ?: run {
+                } else {
                     logger.info("No manga found to fix url with name : ${update.mangaUrl}")
                 }
             }
@@ -105,8 +109,8 @@ class Collector(val luRepo: LatestUpdateRepository,
     }
 
     fun brokenChapterCollector() {
-        val page = 1;
-        val pageItem = 10;
+        val page = 1
+        val pageItem = 10
         var mangas = mangaRepository.findAll(PageRequest.of(page, pageItem))
         val nameSet = HashSet<String>()
         logger.info("Performing batch chapter fixes for ${mangas.totalPages} pages for ${mangas.totalElements} items")
@@ -153,18 +157,19 @@ class Collector(val luRepo: LatestUpdateRepository,
                 var coverImageUrl: String
                 var viewCount: Long
                 elements.forEach {
-                    name = it.select("h3").text();
+                    name = it.select("h3").text()
                     mangaUrl = it.select("a[href]").first().attr("abs:href")
                     description = it.select("p").text()
-                    name = it.select("h3").text();
-                    viewCount = it.select(".aye_icon").text().replace(Regex(","), "")?.toLong()
-                    val link = it.select(".list-story-item-wrap-chapter").select("a[href]").first();
+                    name = it.select("h3").text()
+                    viewCount = it.select(".aye_icon").text().replace(Regex(","), "").toLong()
+                    val link = it.select(".list-story-item-wrap-chapter").select("a[href]").first()
                     latestChapterUrl = link.attr("abs:href")
                     coverImageUrl = it.select("img").first().attr("abs:src")
-
+                    latestChapterUrl = latestChapterUrl.replace(Regex("(s\\d+\\.m)"), "s8.m").replace(Regex("(\\d+\\.com)"), "8.com")
+                    latestChapterUrl = latestChapterUrl.replace("https://bu.mkklcdnbuv8.com", "https://bu.mkklcdnbuv1.com")
                     val mangas = luRepo.findByMangaUrl(mangaUrl)
                     if (mangas.size > 0) {
-                        mangas?.first()?.let { mangaUpdate ->
+                        mangas.first().let { mangaUpdate ->
                             mangaUpdate.mangaName = name
                             mangaUpdate.description = description
                             mangaUpdate.mangaUrl = mangaUrl
@@ -210,8 +215,8 @@ class Collector(val luRepo: LatestUpdateRepository,
 
     fun mangaDetailsCollector() {
         logger.info("Running manga details collector")
-        val page = 1;
-        val pageItem = 20;
+        val page = 1
+        val pageItem = 20
         var mangas = luRepo.findByChapterUpdatedTrue(PageRequest.of(page, pageItem))
         logger.info("Performing batch manga collecton for ${mangas.totalPages} pages for ${mangas.totalElements} items")
         for (x in 1..mangas.totalPages) {
@@ -219,7 +224,7 @@ class Collector(val luRepo: LatestUpdateRepository,
             mangas = luRepo.findByChapterUpdatedTrue(PageRequest.of(x, pageItem))
             mangas.content.forEach {
                 try {
-                    collectMangaDetail(it);
+                    collectMangaDetail(it)
                     it.chapterUpdated = false
                     luRepo.save(it)
                 } catch (e: Exception) {
@@ -246,7 +251,7 @@ class Collector(val luRepo: LatestUpdateRepository,
         if (latestManga.mangaUrl.startsWith("https://mangakakalot.com/")) {
             cover = doc.select(".manga-info-pic").select("img").first().attr("abs:src")
             val elements = doc.select(".manga-info-text").select("li")
-            description = doc.select("#noidungm").text();
+            description = doc.select("#noidungm").text()
             elements?.forEach {
                 if (it.text().contains("Author")) {
                     val authorLinks = it.select("a[href]")
@@ -255,10 +260,10 @@ class Collector(val luRepo: LatestUpdateRepository,
                     }
                 } else if (it.text().contains("Status")) {
                     val statusArr = it.text().split(":")
-                    status = statusArr[1]?.trim()
+                    status = statusArr[1].trim()
                 } else if (it.text().contains("Last updated")) {
                     val updateArr = it.text().split(":")
-                    mangaLastUpdated = updateArr[1]?.trim()
+                    mangaLastUpdated = updateArr[1].trim()
                 } else if (it.text().contains("Genres")) {
                     val genresLink = it.select("a[href]")
                     genresLink?.forEach { genre ->
@@ -273,7 +278,7 @@ class Collector(val luRepo: LatestUpdateRepository,
                 0.0
             }
             doc.select(".chapter-list").select("a[href]")
-            val chapterLinks = doc.select(".chapter-list").select("a[href]");
+            val chapterLinks = doc.select(".chapter-list").select("a[href]")
             chapterLinks?.forEach {
                 val chapterLink = it.attr("abs:href")
                 val chapterName = it.text()
@@ -320,8 +325,8 @@ class Collector(val luRepo: LatestUpdateRepository,
 
         } else if (latestManga.mangaUrl.startsWith("https://manganelo.com/")) {
             cover = doc.select(".info-image").select("img").first().attr("abs:src")
-            description = doc.select(".panel-story-info-description").text();
-            val upperTable = doc.select(".variations-tableInfo").select(".table-value");
+            description = doc.select(".panel-story-info-description").text()
+            val upperTable = doc.select(".variations-tableInfo").select(".table-value")
             upperTable?.let {
                 val authorLinks = it[1].select("a[href]")
                 authorLinks?.forEach { author ->
@@ -436,12 +441,12 @@ class Collector(val luRepo: LatestUpdateRepository,
     }
 
     fun timeElapsedToExecute(function: () -> Unit) {
-        val start = Date();
+        val start = Date()
         logger.info("Started $start")
         function.invoke()
-        val end = Date();
+        val end = Date()
         logger.info("Finished $end")
-        val diffInMillisec: Long = end.getTime() - start.getTime()
+        val diffInMillisec: Long = end.time - start.time
         val diffInMin = TimeUnit.MILLISECONDS.toMinutes(diffInMillisec)
         val diffInSec = TimeUnit.MILLISECONDS.toSeconds(diffInMillisec)
         logger.info("Time took to finish $diffInSec seconds or  $diffInMin minutes")
@@ -451,10 +456,10 @@ class Collector(val luRepo: LatestUpdateRepository,
         logger.info("Start collecting trending manga")
         resetTrendingTag()
         val doc = Jsoup.connect("https://manganelo.com/").timeout(300000).get()
-        val trendingDoc = doc.select(".owl-carousel").select(".slide-caption");
+        val trendingDoc = doc.select(".owl-carousel").select(".slide-caption")
         trendingDoc?.forEach { trend ->
             val url = trend.select(".a-h").first().select("a").attr("href")
-            mangaRepository.findByMangaUrl(url)?.let {
+            mangaRepository.findByMangaUrl(url)?.first()?.let {
                 it.trending = true
                 mangaRepository.save(it)
             }
@@ -466,12 +471,12 @@ class Collector(val luRepo: LatestUpdateRepository,
 
     fun collectMostPopular() {
         logger.info("Start collecting most popular manga")
-        resetMostPopularTag();
+        resetMostPopularTag()
         val doc = Jsoup.connect("https://mangakakalot.com/").timeout(300000).get()
         val mostPopularDoc = doc.select(".owl-carousel").select(".slide-caption")
         mostPopularDoc?.forEach { popular ->
             val url = popular.select("a[href]").first().select("a").attr("href")
-            mangaRepository.findByMangaUrl(url)?.let {
+            mangaRepository.findByMangaUrl(url)?.first()?.let {
                 it.mostPopular = true
                 mangaRepository.save(it)
             }
@@ -481,13 +486,13 @@ class Collector(val luRepo: LatestUpdateRepository,
 
 
     fun resetTrendingTag() {
-        val page = 1;
-        val pageItem = 10;
-        var resetContent = mangaRepository.findByTrendingTrue(PageRequest.of(page, pageItem));
+        val page = 1
+        val pageItem = 10
+        var resetContent = mangaRepository.findByTrendingTrue(PageRequest.of(page, pageItem))
         logger.info("Performing resetting trending status for ${resetContent.totalPages} pages for ${resetContent.totalElements} items")
         for (x in 1..resetContent.totalPages) {
             logger.info("Performing resetting popular status for batch : $x")
-            resetContent = mangaRepository.findByTrendingTrue(PageRequest.of(page, pageItem));
+            resetContent = mangaRepository.findByTrendingTrue(PageRequest.of(page, pageItem))
             resetContent.content.map { mangaCompact -> mangaCompact.trending = false }
             mangaRepository.saveAll(resetContent)
         }
@@ -496,17 +501,45 @@ class Collector(val luRepo: LatestUpdateRepository,
     }
 
     fun resetMostPopularTag() {
-        val page = 1;
-        val pageItem = 10;
-        var resetContent = mangaRepository.findByMostPopularTrue(PageRequest.of(page, pageItem));
+        val page = 1
+        val pageItem = 10
+        var resetContent = mangaRepository.findByMostPopularTrue(PageRequest.of(page, pageItem))
         logger.info("Performing resetting popular status for ${resetContent.totalPages} pages for ${resetContent.totalElements} items")
         for (x in 1..resetContent.totalPages) {
             logger.info("Performing resetting popular status for batch : $x")
-            resetContent = mangaRepository.findByMostPopularTrue(PageRequest.of(x, pageItem));
+            resetContent = mangaRepository.findByMostPopularTrue(PageRequest.of(x, pageItem))
             resetContent.content.map { mangaCompact -> mangaCompact.trending = false }
             mangaRepository.saveAll(resetContent)
         }
         logger.info("Finished resetting popular status for ${resetContent.totalPages} pages for ${resetContent.totalElements} items")
 
+    }
+
+    fun resetLatestChapter() {
+        val page = 1
+        val pageItem = 50
+        var resetContent = luRepo.findAll(PageRequest.of(page, pageItem))
+        logger.info("Performing resetting for latest chapter status for ${resetContent.totalPages} pages for ${resetContent.totalElements} items")
+        for (x in 1..resetContent.totalPages) {
+            logger.info("Performing resetting for latest chapter status for for batch : $x")
+            resetContent = luRepo.findAll(PageRequest.of(x, pageItem))
+            resetContent.content.map { update -> getChapterUpdateStatus(update) }
+            luRepo.saveAll(resetContent)
+        }
+        logger.info("Finished resetting for latest chapter status for ${resetContent.totalPages} pages for ${resetContent.totalElements} items")
+    }
+
+    private fun getChapterUpdateStatus(update: LatestMangaUpdate) {
+        val mangas = mangaCompactRepository.findByMangaUrl(update.mangaUrl)
+        if (mangas == null) {
+            update.chapterUpdated = true
+        } else {
+            mangas.first().let { manga ->
+                update.chapterUpdated = manga.latestChapter != update.latestChapter
+                if (update.chapterUpdated) {
+                    logger.info("${manga.latestChapter} -> ${update.latestChapter}")
+                }
+            }
+        }
     }
 }
